Moves configs around.
