+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 dist
-test
+/test
 .netlify
 .cache
 _generated*
